requirements.txt
sentence-transformers==2.3.1
pinecone-client==4.1.0
streamlit==1.35.0